requirements.txt
Pillow>=11.0.0
PyQt6>=6.7.0
pyinstaller>=6.15.0
pytest>=8.0.0
